backend/requirements.txt
fastapi
uvicorn
python-dotenv
openai
requests
supabase
python-multipart 